moved favorites related classes to new package
