+++ backend/src/main/java/com/cosmic/tracker/favorites/FavoriteController.java
@@ -1,5 +1,7 @@
-package com.cosmic.tracker.Users;
+package com.cosmic.tracker.favorites;
 
+import com.cosmic.tracker.Users.MyUser;
+import com.cosmic.tracker.Users.UserRepository;
 import com.cosmic.tracker.cosmicevent.CosmicEvent;
 import com.cosmic.tracker.cosmicevent.CosmicEventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
